Critters Finally

+++ src/critter/Brain.java
@@ -5,6 +5,10 @@ import java.util.List;
 
 public interface Brain
 {
-	public List<Double> readOutput (float tpf_);
-	public void activate (List<Double> input_, float tpf_);
+	public List<Float> readOutput (float tpf_);
+	public void activate (List<Float> input_, float tpf_);
+
+	// These should be static
+//	public int nnInputs (int inputs_);
+//	public int nnOutputs (int outputs_);
 }
